requirements.txt
anthropic>=0.18.0
google-cloud-bigquery>=3.14.0
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.38.0
PyGithub>=2.1.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
